rename seedance2pro crate -> seedance2pro_client
