+++ crates/api_clients/artcraft/artcraft_router/Cargo.toml
@@ -20,7 +20,7 @@ artcraft_api_defs.workspace = true
 artcraft_client.workspace = true
 fal_client.workspace = true
 muapi_client.workspace = true
-seedance2pro.workspace = true
+seedance2pro_client.workspace = true
 tokens.workspace = true
 url_utils.workspace = true
 world_labs_api.workspace = true
