+++ crates/api_clients/seedance2pro_client/src/error/seedance2pro_specific_api_error.rs
