+++ crates/api_clients/seedance2pro_client/src/requests/prepare_file_upload/request_types.rs
