+++ crates/service/job/seedance2_pro_job/src/job_dependencies.rs
@@ -1,7 +1,7 @@
 use cloud_storage::bucket_client::BucketClient;
 use concurrency::relaxed_atomic_bool::RelaxedAtomicBool;
 use jobs_common::job_stats::JobStats;
-use seedance2pro::creds::seedance2pro_session::Seedance2ProSession;
+use seedance2pro_client::creds::seedance2pro_session::Seedance2ProSession;
 use server_environment::ServerEnvironment;
 use sqlx::MySqlPool;
