+++ crates/service/job/seedance2_pro_job/src/process_job/process_successful_job.rs
@@ -12,7 +12,7 @@ use hashing::sha256::sha256_hash_bytes::sha256_hash_bytes;
 use mysql_queries::queries::generic_inference::seedance2pro::list_pending_seedance2pro_jobs::PendingSeedance2ProJob;
 use mysql_queries::queries::media_files::create::insert_builder::media_file_insert_builder::MediaFileInsertBuilder;
 use mysql_queries::queries::generic_inference::web::mark_generic_inference_job_successfully_done_by_token::mark_generic_inference_job_successfully_done_by_token;
-use seedance2pro::requests::poll_orders::poll_orders::OrderStatus;
+use seedance2pro_client::requests::poll_orders::poll_orders::OrderStatus;
 use crate::job_dependencies::JobDependencies;
 
 const PREFIX: &str = "artcraft_";
